// src/app/shared/model/notification-h-obj.model.ts
export class NotificationHObj {
    NotificationHId: any;
    MrNotificationTypeCode: any;
    MrNotificationGroupCode: any;
    MrNotificationMethodCode: any;
    Title: any;
    ShortDesc: any;
    LongDesc: any;
    ReqBy: any;
    ApproveBy: any;
    PublishDt: any;
    Status: any;
    IsActive: any;
    IsDraft: any;
    listTargetRefOfficeId = [];
    listTargetRefRoleId = [];
    listUsername = [];
    RowVersion: any;
    constructor() { this.NotificationHId = 0, this.RowVersion = "" }
  }
  